正式建立256*256的缓存表，加快较大字符串的查询速度；对场景 #30 比较有效 close  #30

+++ unit_test.cpp
@@ -167,7 +167,7 @@ static inline long test(const char* newStr,const char* oldStr,const char* error_
 
 
 void setRandDataSize(int kMaxDataSize,std::vector<TByte>& oldData,std::vector<TByte>& newData,
-                            float sunSizeMin,float subSizeMax){
+                            double sunSizeMin,double subSizeMax){
     const TInt oldSize=(TInt)(rand()*(1.0/RAND_MAX)*rand()*(1.0/RAND_MAX)*kMaxDataSize);
     const TInt newSize=(TInt)(oldSize*(sunSizeMin+rand()*(1.0/RAND_MAX)*(subSizeMax-sunSizeMin)));
     newData.resize(newSize);
